+++ tests/core/filtering/conftest.py
@@ -168,7 +168,7 @@ def return_filter(contract=None, args=[]):
     kwargs = apply_key_map({"filter": "argument_filters"}, args[1])
     if "fromBlock" not in kwargs:
         kwargs["fromBlock"] = "latest"
-    return contract.events[event_name].createFilter(**kwargs)
+    return contract.events[event_name].create_filter(**kwargs)
 
 
 @pytest.fixture(scope="module")