+++ web3/contract.py
@@ -1474,7 +1474,7 @@ class BaseContractEvent:
         return get_event_data(self.w3.codec, self.abi, log)
 
     @combomethod
-    def createFilter(
+    def create_filter(
         self,
         *,  # PEP 3102
         argument_filters: Optional[Dict[str, Any]] = None,
@@ -1489,7 +1489,7 @@ class BaseContractEvent:
         """
         if fromBlock is None:
             raise TypeError(
-                "Missing mandatory keyword argument to createFilter: fromBlock"
+                "Missing mandatory keyword argument to create_filter: fromBlock"
             )
 
         if argument_filters is None:
@@ -1615,7 +1615,7 @@ class ContractEvent(BaseContractEvent):
     ) -> Iterable[EventData]:
         """Get events for this contract instance using eth_getLogs API.
 
-        This is a stateless method, as opposed to createFilter.
+        This is a stateless method, as opposed to create_filter.
         It can be safely called against nodes which do not provide
         eth_newFilter API, like Infura nodes.
 
@@ -1690,7 +1690,7 @@ class AsyncContractEvent(BaseContractEvent):
     ) -> Awaitable[Iterable[EventData]]:
         """Get events for this contract instance using eth_getLogs API.
 
-        This is a stateless method, as opposed to createFilter.
+        This is a stateless method, as opposed to create_filter.
         It can be safely called against nodes which do not provide
         eth_newFilter API, like Infura nodes.
 
@@ -1952,12 +1952,12 @@ def check_for_forbidden_api_filter_arguments(
         _input = name_indexed_inputs[filter_name]
         if is_array_type(_input["type"]):
             raise TypeError(
-                "createFilter no longer supports array type filter arguments. "
+                "create_filter no longer supports array type filter arguments. "
                 "see the build_filter method for filtering array type filters."
             )
         if is_list_like(filter_value) and is_dynamic_sized_type(_input["type"]):
             raise TypeError(
-                "createFilter no longer supports setting filter argument options for "
+                "create_filter no longer supports setting filter argument options for "
                 "dynamic sized types. See the build_filter method for setting "
                 "filters with the match_any method."
             )